# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SolomonQuest LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users in the school
      parameters:
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Change a user's role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # ─── PROGRAMS ───────────────────────────────────────────────────────────────
...
  /assignments/{assignmentId}/submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions for an assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: submitAssignment
      tags: [submissions]
      summary: Submit an assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/grade:
    patch:
      operationId: gradeSubmission
      tags: [submissions]
      summary: Grade a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "200":
          description: Graded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # ─── ANNOUNCEMENTS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id]
      properties:
        id:
          type: string
        schoolId:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
    ProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string

    # ─── PROGRAM ─────────────────────────────────────────────────────────────
...
    Submission:
      type: object
      required: [id, assignmentId, studentId, status]
      properties:
        id:
          type: string
        assignmentId:
          type: string
        studentId:
          type: string
        content:
          type: ["string", "null"]
        grade:
          type: ["number", "null"]
        status:
          type: string
        studentName:
          type: ["string", "null"]
    SubmissionInput:
      type: object
      properties:
        content:
          type: string
    GradeInput:
      type: object
      required: [grade]
      properties:
        grade:
          type: number
        feedback:
          type: string

    # ─── ANNOUNCEMENT ────────────────────────────────────────────────────────
...